Pebble Hills University
Global Training Academy and Pebble Hills University provide an amazing opportunity for both professional and aspiring teachers to embark on an extraordinary TESOL career by introducing an exclusive and flexible program-Master of Arts in Teaching English to Speakers of Other Languages. The program not only equips the teachers with current language teaching methods and varied strategies of instructions but provides a framework for a fascinating career ahead. The world of TESOL is constantly changing and this internationally recognized program is designed in a way so that the trainees can gain specialized training to carry out the teaching activities in a productive and professional manner.

http://www.pebblehills.edu